Add Feature: Create Indexing
After preprocessing, datafiles will be indexed and store in Lucene
RAMDirectory and wait for the search query.

+++ src/xaurora/dropbox/DropboxAuth.java
@@ -5,11 +5,6 @@ import java.util.Locale;
 
 import xaurora.util.UserProfile;
 
-import com.dropbox.core.DbxAccountInfo;
-import com.dropbox.core.DbxClient;
-import com.dropbox.core.DbxException;
-import com.dropbox.core.DbxRequestConfig;
-
 public class DropboxAuth {
 	//unclear declaration
 	final static String APP_KEY = "4tpptik431fwlqo";
@@ -96,7 +91,7 @@ public class DropboxAuth {
 	
 	protected static String bytesToGB(long quota){		
 		//Magic String
-		double GB = (double) quota / (1024.0*1024.0*1024.0);
+		double GB = quota / (1024.0*1024.0*1024.0);
 		DecimalFormat df = new DecimalFormat("#.##");
 		//Magic String
 		return df.format(GB);
